requirements.txt
streamlit>=1.32.0
openpyxl>=3.1.2
pandas>=2.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
